Moved plugs to own directory and reanmed to better express intent in preparation for new cookie verification plug.
