+++ README.md
@@ -54,10 +54,10 @@ end
 Two plugs are provided:
 
 ```
-- Jwt.Plugs.VerifySignature
+- Jwt.Plugs.VerifyAuthorizationHeader
 ```
 
-The plug looks at the authorization HTTP header to see if it includes a value with the format
+The plug looks at the HTTP Authorization header to see if it includes a value with the format
 
 ```
 Authorization: Bearer [JWT]
